Move src/bin/tokenizer_test.rs to tests/tokenizer.rs
