+++ tests/lib.rs
@@ -1,0 +1,4 @@
+extern crate regex;
+extern crate serde_derive;
+
+mod tokenizer;
